requirements.txt
aniso8601==7.0.0
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
choicesenum==0.7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-bootstrap4==2.3.1
django-exiffield==3.0.0
django-extensions==3.0.9
django-filter==2.4.0
django-graphiql==0.4.4
django-graphql-auth==0.3.14
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-location-field==2.1.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.2
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonfield==3.1.0
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
promise==2.3
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
pydot==1.4.1
Pygments==2.7.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.18.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.1
tornado==6.0.4
traitlets==5.0.5
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
